# smart-backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.CreateUserRequest:
    properties:
      locations:
        description: Optional - if provided, will assign user to these locations and
          gates
        items:
          $ref: '#/definitions/handlers.LocationAssignmentRequest'
        type: array
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "+77771234567"
        type: string
    required:
    - password
    - phone
    type: object
...
  handlers.GateDTO:
    properties:
      description:
        example: Main vehicle entrance for visitors. Controlled by biometric access,
          opens in 3 seconds with safety sensors.
        type: string
      gate_is_horizontal:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      is_open:
        example: true
        type: boolean
      location_id:
        example: 1
        type: integer
      title:
        example: Автоматический Шлагбаум №12
        type: string
    type: object
...
  handlers.LocationAssignmentRequest:
    properties:
      gateIds:
        items:
          type: integer
        type: array
      locationId:
        example: 1
        type: integer
    required:
    - gateIds
    - locationId
    type: object
  handlers.LocationDTO:
    properties:
      address:
        example: г. Бишкек, проспект Чуй, 135
        type: string
      gates:
        description: Always include gates, even if empty array
        items:
          $ref: '#/definitions/handlers.GateDTO'
        type: array
      id:
        example: 1
        type: integer
      logo:
        example: https://picsum.photos/seed/alatoo/200
        type: string
      title:
        example: Торгово-развлекательный центр Ала-Тоо
        type: string
    type: object
...
  handlers.PaginationMeta:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 1
        type: integer
      per_page:
        example: 100
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      locations:
        description: Optional - if provided, will reassign user to these locations
          and gates
        items:
          $ref: '#/definitions/handlers.LocationAssignmentRequest'
        type: array
      password:
        description: Optional - only updates if provided
        example: newpassword123
        minLength: 6
        type: string
      phone:
        description: Optional - if provided, will update phone number after checking
          availability
        example: "+77771234567"
        type: string
    type: object
  handlers.UserDTO:
    properties:
      created_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+77771234567"
        type: string
      updated_at:
        example: "2025-01-15T10:30:00Z"
        type: string
    required:
    - created_at
    - id
    - phone
    - updated_at
    type: object
  handlers.UserData:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+77771234567"
        type: string
    required:
    - id
    - phone
    type: object
  handlers.UserDetailDTO:
    properties:
      created_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      locations:
        items:
          $ref: '#/definitions/handlers.LocationDTO'
        type: array
      phone:
        example: "+77771234567"
        type: string
      updated_at:
        example: "2025-01-15T10:30:00Z"
        type: string
    required:
    - created_at
    - id
    - locations
    - phone
    - updated_at
    type: object
  handlers.UserDetailResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.UserDetailDTO'
      message:
        example: User retrieved successfully
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
  handlers.UserResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.UserData'
      message:
        example: User created successfully
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
  handlers.UsersListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.UserDTO'
        type: array
      message:
        example: Users retrieved successfully
        type: string
      pagination:
        $ref: '#/definitions/handlers.PaginationMeta'
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ololo-gate.com
    name: API Support
  description: Secure phone-based authentication backend for Ololo Gate management
    system with dual authentication (users & admins) and role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ololo Gate API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all registered users with pagination and search
        (requires admin authentication)
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Records per page (default: 500)'
        in: query
        name: limit
        type: integer
      - description: Search by phone number
        in: query
        name: search
        type: string
      - description: 'Order results by created_at (ASC or DESC, default: DESC)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/handlers.UsersListResponse'
        "401":
          description: Unauthorized - invalid or missing admin token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Create a new user account and assign locations and gates via third-party
        API (requires admin authentication)
      parameters:
      - description: User creation details with locations and gates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized - invalid or missing admin token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "409":
          description: User with this phone number already exists
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error or third-party API failure
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new user with location and gate assignment
      tags:
      - User Management
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user account by ID (soft delete, requires admin authentication)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized - invalid or missing admin token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Retrieve a specific user's details by ID including their assigned
        locations and gates from third-party API (requires admin authentication)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully with locations
          schema:
            $ref: '#/definitions/handlers.UserDetailResponse'
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized - invalid or missing admin token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID with assigned locations and gates
      tags:
      - User Management
    patch:
      consumes:
      - application/json
      description: Update a user's password (optional) and reassign locations and
        gates via third-party API (requires admin authentication)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update details (password optional, locations and gates required)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Invalid user ID or request body
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized - invalid or missing admin token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error or third-party API failure
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user password and location/gate assignments
      tags:
      - User Management
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"